requirements.txt
flask==3.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
python-dotenv==1.0.0
ta==0.11.0
matplotlib==3.8.2
plotly==5.18.0
krakenex==2.2.0
ccxt==4.2.0
